auth: move login task types
